==> Training_Income_v1.3/1-SetUp.ipynb <==
# Lab 1 Setup and Test 
In this lab we will:
1. introduce and test the line magic functions that are apart of Ezmeral MLOps
2. populate our project repository with the necessary files 

<b>Line magics</b>, which are denoted by a single % prefix, operate on a single line of input.

<b>Cell magics</b>, which are denoted by a double %% prefix, operate on multiple lines of input.

# Magics

#### **Get the list of connected training environments**

<b>%attachments</b> is a line magic command that output a table with the name(s) of the training cluster(s) available for us to use. Sometimes, Operations team may have created multiple training clusters for different projects depending on the needs of the model or size of data, e.g. some with GPU nodes, while others with CPUs only.

%attachments

#### **Submit a notebook code cell to a remote training cluster**

To utilize the training cluster, you will need grab the name of the training cluster you want to use and feed it into another custom cell magic command. 

Assuming your training cluster name is called "training", using **%%training** magic command specified at the beginning of the cell will tell the Jupyter Notebook to
submit the entire content of the cell to the training cluster named _training_. If you comment out this magic command, the code will run on your local Jupyter Notebook.

The example cell below will execute a print statement on the training cluster.

%%<your training cluster>
print("Hello world")

#### **Retrieve the result of the job**

The training cluster will send back a unique log url for the job submitted.   
You can use the _History URL_ with the _"%log --url"_ custom **line magic** command to track the status of the job in real time. 

Copy the History URL from the output of the previous cell and paste it into the cell below where it says _"your_http_url_here"_, and run the cell code.

A status of "**Finished**" means the execution of the job submitted to the training cluster has completed.

# Continue onto Lab 2 for some data engineering work!

# Fill in the history url from the output of the previous cell
%logs --url <your history url>

# Add your initials to the XGB_Scoring_K8s.py file on line 12

# Setup your working directory

The Python code here is used to create the working directory in the central project repository for your user. The working directory is used to store the key data components needed in the ML pipeline, such as the input dataset, trained ML model(s) and scoring script(s). 

Fill in your initials for the <b>STUDENT</b> variable 

import os
import shutil

# Fill in your initials here
STUDENT = ""
%store STUDENT

usecaseDirectory = STUDENT + "_UCI_Income"
RepoModel = "models" + '/' + usecaseDirectory
RepoCode = "code" + '/' + usecaseDirectory
RepoData = "data" + '/' + usecaseDirectory

# Project repo path function - file system mount available to all app containers
def ProjectRepo(path):
    ProjectRepo = "/bd-fs-mnt/project_repo"
    return str(ProjectRepo + '/' + path)

print("Creating your directories..")
if (not os.path.exists(ProjectRepo(RepoModel))):
    print ("Creating the Model directory " + RepoModel)
    try:
        os.makedirs(ProjectRepo(RepoModel))
    except OSError:
        print ("Creation of the model directory %s failed" % RepoModel)
    else:
        print ("Successfully created the model directory %s" % RepoModel)

if (not os.path.exists(ProjectRepo(RepoCode))):
    print ("Creating the Code directory " + RepoCode)
    try:
        os.makedirs(ProjectRepo(RepoCode))
    except OSError:
        print ("Creation of the code directory %s failed" % RepoCode)
    else:
        print ("Successfully created the code directory %s" % RepoCode)
        
if (not os.path.exists(ProjectRepo(RepoData))):
    print ("Creating the Data directory " + RepoData)
    try:
        os.makedirs(ProjectRepo(RepoData))
    except OSError:
        print ("Creation of the code directory %s failed" % RepoData)
    else:
        print ("Successfully created the code directory %s" % RepoData)
        
print("Loading files into project repo")
shutil.copy("Data/adult_data.csv", ProjectRepo(RepoData))
shutil.copy("Data/adult_test.csv", ProjectRepo(RepoData))
shutil.copy("XGB_Scoring_K8s.py", ProjectRepo(RepoCode))

## View ProjectRepo 
- Notice the files and that the permissions here match that of the AD/LDAP user  

!ls -ltr /bd-fs-mnt/project_repo/data

### Now return back to the ECP UI and browse the project repository to view your files!
